+++ src/TitleDetails/AddPieceModal/AddPieceModalContainer.js
@@ -46,6 +46,7 @@ const AddPieceModalContainer = ({
       onCheckIn={onCheckIn}
       onSubmit={onSubmit}
       pieceFormatOptions={pieceFormatOptions}
+      poLine={poLine}
     />
   );
 };