+++ src/common/components/index.js
@@ -1,1 +1,2 @@
 export * from './CreateItemField';
+export * from './LineLocationsView';