+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Allow user to adjust order quantity and location/hodlings connection from receiving ui. Refs UIREC-132.
 * Save piece toast message message. Refs UIREC-159.
 * Increment stripes to v7. Refs UIREC-154.
+* Display order line locations on piece form. Refs UIREC-163.
 
 ## [1.4.2](https://github.com/folio-org/ui-receiving/tree/v1.4.2) (2021-08-05)
 [Full Changelog](https://github.com/folio-org/ui-receiving/compare/v1.4.1...v1.4.2)